Add files via upload
The logistic regression not only gives a measure of how relevant a predictor is (coefficient size) but also its direction of association (positive or negative). The output of a logistic regression is more informative than other classification algorithms. Like any regression approach, it expresses the relationship between an outcome variable (label) and each of its predictors (features).